+++ src/main/data/softess.txt
@@ -1,2 +1,3 @@
 TODO|0| walk dog
 DEADLINE|1| walker | 2024-12-22 1234
+DEADLINE|0| watch movie | 2024-12-23 2212
