+++ src/main/java/softess/Command.java
@@ -1,3 +1,5 @@
+package softess;
+
 public abstract class Command {
