+++ src/main/java/softess/DataHandler.java
@@ -1,3 +1,5 @@
+package softess;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
