+++ src/main/java/softess/DeleteCommand.java
@@ -1,3 +1,5 @@
+package softess;
+
 public class DeleteCommand extends Command {
 
     public TaskList tasks;
