+++ src/main/java/softess/ExitCommand.java
@@ -1,3 +1,5 @@
+package softess;
+
 public class ExitCommand extends Command {
 
     public ExitCommand(UserInterface ui) {
