+++ src/main/java/softess/MarkCommand.java
@@ -1,3 +1,5 @@
+package softess;
+
 public class MarkCommand extends Command {
 
     public TaskList tasks;
