+++ src/main/java/softess/Softess.java
@@ -1,3 +1,5 @@
+package softess;
+
 import java.io.IOException;
 import java.util.Scanner;
