+++ src/main/java/softess/Task.java
@@ -1,3 +1,5 @@
+package softess;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
