+++ src/main/java/softess/ToDo.java
@@ -1,18 +1,16 @@
-public class Event extends Task {
-    protected String from;
-    protected String to;
+package softess;
 
-    public Event(String description, String from, String to, boolean isDone) {
+public class ToDo extends Task {
+
+
+    public ToDo(String description, boolean isDone) {
         super(description);
         super.isDone = isDone;
-        this.from = from;
-        this.to = to;
-
     }
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
+        return "[T]" + super.toString();
     }
 
     @Override
@@ -20,15 +18,17 @@ public class Event extends Task {
         super.isDone = true;
         System.out.println("Nice! I've marked this task as done: \n " + this.toString());
     }
+
     @Override
     public void markAsUnDone() {
         super.isDone = false;
         System.out.println("Nice! I've marked this task as undone: \n " + this.toString());
     }
+
     @Override
     public String generateTextToFile() {
         int status = this.isDone ? 1 : 0;
-        String result = "EVENT|%d|%s|%s|%s".formatted(status,this.description,this.from,this.to);
+        String result = "TODO|%d|%s".formatted(status,this.description);
         return result;
     }
 }
