+++ src/main/java/softess/UserInterface.java
@@ -1,14 +1,16 @@
+package softess;
+
 public class UserInterface {
 
     public UserInterface() {}
 
     public void showWelcomeMessage() {
-        System.out.println("Hello! I'm Softess");
+        System.out.println("Hello! I'm softess.Softess");
         System.out.println("What can I do for you?\n");
     }
 
     public void showGoodByeMessage() {
-        System.out.println("Softess: Bye. Hope to see you again soon noob!");
+        System.out.println("softess.Softess: Bye. Hope to see you again soon noob!");
     }
 
     public void showErrorMessage(String message) {
