+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Softess
+Hello! I'm softess.Softess
 What can I do for you?
 
 Got it. I've added this task:
